refactor: extract API business logic into dedicated service modules
This commit extracts complex query building, LLM orchestration, and moderation logic out of route handlers and into testable, dependency-injected services. Implemented using TDD Red-Green-Refactor.
