+++ src/lib/error-handler.ts
@@ -13,21 +13,23 @@ export class ApiError extends Error {
     constructor(
         public statusCode: number,
         message: string,
-        public details?: unknown
+        public details?: unknown,
+        public code?: string
     ) {
         super(message);
         this.name = "ApiError";
     }
 }
 
-export function createApiError(statusCode: number, message: string, details?: unknown): ApiError {
-    return new ApiError(statusCode, message, details);
+export function createApiError(statusCode: number, message: string, details?: unknown, code?: string): ApiError {
+    return new ApiError(statusCode, message, details, code);
 }
 
 interface SanitisedError {
     statusCode: number;
     message: string;
     details?: unknown;
+    code?: string;
     stack?: string[];
 }
 
@@ -36,6 +38,7 @@ export function sanitizeError(error: unknown): SanitisedError {
         return {
             statusCode: error.statusCode,
             message: isProduction ? "An internal error occurred." : error.message,
+            code: error.code,
             ...(isProduction ? {} : { details: error.details }),
             ...(isProduction ? {} : error.stack ? { stack: error.stack.split("\n") } : {}),
         };
@@ -59,6 +62,7 @@ export function buildErrorResponse(error: unknown) {
     const result = sanitizeError(error);
     const body: Record<string, unknown> = { error: result.message };
     if (result.details !== undefined) body.details = result.details;
+    if (result.code !== undefined) body.code = result.code;
     if (result.stack !== undefined) body.stack = result.stack;
     const status = result.statusCode;
     return { status, body } as { status: number; body: Record<string, unknown> };
