+++ src/services/room.service.ts
@@ -1,0 +1,3 @@
+import { ServiceError } from "@/lib/errors";
+
+// TODO
